Remove dependency on bank module for staking and account
