+++ x/account/alias.go
@@ -1,0 +1,9 @@
+package account
+
+import "github.com/TruStory/truchain/x/bank/exported"
+
+const (
+	TransactionRegistration = exported.TransactionRegistration
+	TransactionBacking = exported.TransactionBacking
+
+)
